+++ pkg/monitoring/alloy/service.go
@@ -3,6 +3,7 @@ package alloy
 import (
 	"context"
 	_ "embed"
+	"slices"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
@@ -12,8 +13,8 @@ import (
 
 	"github.com/pkg/errors"
 
+	"github.com/giantswarm/observability-operator/api/v1alpha1"
 	"github.com/giantswarm/observability-operator/pkg/common"
-	commonmonitoring "github.com/giantswarm/observability-operator/pkg/common/monitoring"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
 	"github.com/giantswarm/observability-operator/pkg/common/password"
 	"github.com/giantswarm/observability-operator/pkg/monitoring"
@@ -38,7 +39,7 @@ func (a *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluste
 
 	// Get list of tenants
 	var tenants []string
-	tenants, err := commonmonitoring.ListTenants(a.Client, ctx)
+	tenants, err := listTenants(ctx, a.Client)
 	if err != nil {
 		return errors.WithStack(err)
 	}
@@ -98,4 +99,28 @@ func (a *Service) ReconcileDelete(ctx context.Context, cluster *clusterv1.Cluste
 
 	logger.Info("alloy-service - ensured alloy is removed")
 	return nil
+}
+
+func listTenants(ctx context.Context, k8sClient client.Client) ([]string, error) {
+	tenants := make([]string, 0)
+	var grafanaOrganizations v1alpha1.GrafanaOrganizationList
+
+	err := k8sClient.List(ctx, &grafanaOrganizations)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, organization := range grafanaOrganizations.Items {
+		if !organization.DeletionTimestamp.IsZero() {
+			continue
+		}
+
+		for _, tenant := range organization.Spec.Tenants {
+			if !slices.Contains(tenants, string(tenant)) {
+				tenants = append(tenants, string(tenant))
+			}
+		}
+	}
+
+	return tenants, nil
 }